moved genesis to parent folder for easy upgrade
